Move tracking classes to subpackage
- Move tracking classes to /track/*.py subpackage
- Split tracker.py into motion.py, tracker.py, tracks.py (by class)
- Rename *MotionModel classes to *Motion
