+++ src/glimpse/track/__init__.py
@@ -1,0 +1,8 @@
+"""Track image features through time using particle filtering."""
+
+from .motion import CartesianMotion, CylindricalMotion
+from .observer import Observer
+from .tracker import Tracker
+from .tracks import Tracks
+
+__all__ = ["CartesianMotion", "CylindricalMotion", "Observer", "Tracker", "Tracks"]
